Force application to always be English in the test env.
Reason: feature specs failed on my machine. My Firefox defaults to German,
so the page was delivered in German because of `accept_header_lang`. The
tests look for buttons like "Sign Up" and so on, and fail because its
"Einloggen" now.

There doesn't seem to be a way to configure Selenium's browser instance
to send Accept-Language: en, I'd prefer that.
Poltergeist lets you do that, btw.